Rescue same-file near-misses, complete the within-file bucket star, and read literals as parameters where nothing was renamed
Four recall defects, all inside one file.

[FUSED-CANDIDATE-BUCKET-STAR] A structural-hash bucket paired every
member only with the member that sorted first. Inside one file that
member decides every pair, so when it was the one that *differed* the
byte-identical copy behind it was never a candidate at all: one
unrelated sibling deleted an exact duplicate from the report. Members
of a bucket that share a file are now paired with each other, all of
them — recall may not depend on what else happens to share the shape.

[FUSED-SHARED-SUBTREE-SAME-FILE] The shared-subtree rescue was
cross-file only, so two methods that drifted apart inside one file
published the statement fragments they share and never the methods. It
now reaches them on three conditions: both endpoints are whole authored
declarations and disjoint; they still enclose a Merkle-equal clone of
at least the shared-subtree node floor — authored code the edit never
touched; and the shared mass *beyond* that clone clears the same floor,
which is the echo rule turned inward. Shape and agreement cannot
separate a drifted copy from a shape family (csharp-merge-drift
measures overlap 0.84 / agreement 0.55 against dart-issue-197's
0.81-0.88 / up to 0.56). Copied code can.

[FUSED-CONTENT-GATE-PARAMETER] Where the identifier bijection claims no
rename, Baker's prev-encoding reaches the literal alphabet too: a
substitution seen once is an unconstrained wildcard, not a
contradiction. Two declarations keeping every identifier and every call
while substituting at twelve literal positions are one parameterised
method, and judging them on agreement alone judged a literal-only copy
on the one axis its own edit demolishes. A repeated substitution stays
constrained, so a sibling sharing a shape and no byte cannot join the
copy beside it; an inconsistent one stays constrained too.

[CLONE-NOISE-LITERAL-VARIATION-CALLS] A call consumes its receiver, not
just its arguments: expect(generated).toContain(...) consumes
`generated`. Reading the argument list alone let a scenario family
block its own suppression.

rename.rs and rescue.rs had gone past 500 lines and are split.
