+++ crates/deslop-core/src/content/rename.rs
@@ -20,11 +20,14 @@ use std::{collections::BTreeMap, collections::HashMap, hash::BuildHasher};
 use crate::{buckets::CONTENT_SUPPORT_FLOOR, content::PairScope, state::FileId};
 
 use super::{
-    frontier::{
-        frontiers_aligned, leaf_bytes, member_count, population, MemberContent, Population,
-    },
+    frontier::{frontiers_aligned, member_count, population, MemberContent, Population},
     vacuous_share,
 };
+
+use literal_echo::{affirming_literal_count, literal_echoes, LiteralEchoes};
+
+/// Literal echoes of a rename, and the byte transform that proves one.
+mod literal_echo;
 
 /// Minimum occurrences of a substituted identifier pair before it counts
 /// as rename evidence ([TECH-PMATCH-BAKER]). In Baker's prev-encoding a
@@ -60,9 +63,11 @@ const RENAME_EVIDENCE_HALF_MASS: f64 = 4.0;
 /// promote floor's conservatism: a same-file rename family is the #197
 /// sibling shape, and its literal axis must vouch on its own.
 /// The pool opens only where the literal population affirms at all:
-/// aligned literals with zero preservation and zero echoes are the
+/// constrained literals with zero preservation and zero echoes are the
 /// #134 stride family — every substantive byte disagrees and nothing
-/// outside the substitution vouches, so the axis is `0.0`.
+/// outside the substitution vouches, so the axis is `0.0`. Which
+/// literals are constrained is [`LiteralEvidence::measure`]'s call
+/// ([FUSED-CONTENT-GATE-PARAMETER]).
 ///
 /// Baker's prev-encoding is the discriminator the deleted literal-anchor
 /// cliff could not provide: a substituted identifier pair seen once is
@@ -97,33 +102,135 @@ pub(super) fn pair_rename_consistency<S: BuildHasher>(
         &population(&canonical.keys, &member.keys, Population::Identifier),
         &echoes.per_substitution,
     );
-    let literal_total = population(&canonical.keys, &member.keys, Population::Literal).len();
-    let affirming_literals = affirming_literal_count(canonical, member, &echoes);
-    if affirming_literals == 0 && literal_total > 0 {
+    let literals = LiteralEvidence::measure(canonical, member, &echoes, &mapping);
+    if literals.affirming == 0 && literals.constrained > 0 {
         return 0.0;
     }
-    // A window carved from inside a function that carries no literal at
-    // all offers the substitution nothing to contradict — the literal
-    // that would is on the line the window left out — so a substitution
-    // corroborated only by its own repetition cannot anchor it. Its
-    // anchors are the positions the rename did not supply: identity
-    // identifiers ([FUSED-CONTENT-GATE-INTERIOR]). A whole authored
-    // function or module with no literal is judged as before.
-    let identifier_anchors = if scope.interior && literal_total == 0 {
-        mapping.identity
-    } else {
-        mapping.explained
-    };
-    let anchors = affirming_literals.saturating_add(identifier_anchors);
-    let coverage = if scope.same_file {
-        vacuous_share(affirming_literals, literal_total)
-            .min(vacuous_share(mapping.explained, mapping.constrained))
-    } else {
-        let explained = mapping.explained.saturating_add(affirming_literals);
-        let constrained = mapping.constrained.saturating_add(literal_total);
-        vacuous_share(explained, constrained)
-    };
+    let anchors = literals
+        .affirming
+        .saturating_add(mapping.anchors(scope, literals.aligned));
+    let coverage = literals.coverage(&mapping, scope);
     coverage * evidence_weight(coverage, anchors)
+}
+
+/// The pair's aligned literal positions, split into what the coverage
+/// must explain and what it does explain ([FUSED-CONTENT-GATE],
+/// [FUSED-CONTENT-GATE-PARAMETER]).
+struct LiteralEvidence {
+    /// Every aligned literal position, whatever it says.
+    aligned: usize,
+    /// Positions the coverage must explain — see
+    /// [`LiteralEvidence::measure`].
+    constrained: usize,
+    /// Positions that affirm the copy: preserved bytes or an echo of a
+    /// bijection-explained substitution.
+    affirming: usize,
+}
+
+impl LiteralEvidence {
+    /// Measures one pair's literal positions.
+    ///
+    /// A preserved literal and a literal echo affirm the copy at the
+    /// position itself. A drifted literal that echoes nothing
+    /// contradicts the *rename* the identifier bijection claims, so it
+    /// is constrained and unexplained — the `#134` stride family renames
+    /// consistently end to end and diverges at one aligned literal, and
+    /// that one position is the whole difference between it and a
+    /// reportable Type-2 clone.
+    ///
+    /// [FUSED-CONTENT-GATE-PARAMETER] Where the bijection claims no
+    /// rename — no substituted identifier position is corroborated —
+    /// there is no claim for a drifted literal to contradict, and
+    /// [TECH-PMATCH-BAKER]'s prev-encoding applies to the literal
+    /// alphabet exactly as it does to the identifier one: a substitution
+    /// seen *once* is an unconstrained wildcard. Two declarations whose
+    /// every identifier position is byte-identical and whose literals
+    /// each substitute once are one parameterised declaration, and those
+    /// literals are its parameters — `csharp-merge-manyholes` keeps
+    /// every identifier and every call and substitutes at all twelve
+    /// literal positions, which is what `[AUTOFIX-MERGE-GATE]`
+    /// independently calls a clone too parameterised to merge
+    /// mechanically.
+    ///
+    /// A *repeated* substitution is not a wildcard. It is the sibling
+    /// family's own subject carried through its body — the star-shadow
+    /// fixture's `ApplyAlpha` says `"alpha"` three times against
+    /// `"dup"` — and it stays constrained, so a sibling that shares a
+    /// shape and no byte cannot join the copy it sits beside. An
+    /// inconsistent substitution stays constrained too: it contradicts
+    /// the parameterisation as surely as it would a rename.
+    fn measure(
+        canonical: &MemberContent,
+        member: &MemberContent,
+        echoes: &LiteralEchoes,
+        mapping: &RenameMapping,
+    ) -> Self {
+        let positions = literal_positions(canonical, member);
+        let affirming = affirming_literal_count(canonical, member, echoes);
+        let pairs = literal_pairs(&positions);
+        let bijection = ModalBijection::over(&substituted_pairs(&pairs));
+        let occurrences = pair_counts(pairs.iter().copied());
+        let constrained = if mapping.renames() {
+            positions.len()
+        } else {
+            positions
+                .iter()
+                .filter(|(index, keys)| {
+                    keys.0 == keys.1
+                        || echoes.positions.contains(index)
+                        || !bijection.explains(keys)
+                        || occurrences.get(keys).copied().unwrap_or_default()
+                            >= RENAME_CORROBORATION_MIN_OCCURRENCES
+                })
+                .count()
+        };
+        Self {
+            aligned: positions.len(),
+            constrained,
+            affirming,
+        }
+    }
+
+    /// The pooled coverage over this pair's constrained positions.
+    ///
+    /// A cross-file pair pools the literal and identifier populations
+    /// into one share. A same-file pair keeps the stricter min of the
+    /// two, matching the promote floor's conservatism: a same-file
+    /// rename family is the `#197` sibling shape, and its literal axis
+    /// must vouch on its own.
+    fn coverage(&self, mapping: &RenameMapping, scope: PairScope) -> f64 {
+        if scope.same_file {
+            return vacuous_share(self.affirming, self.constrained)
+                .min(vacuous_share(mapping.explained, mapping.constrained));
+        }
+        vacuous_share(
+            mapping.explained.saturating_add(self.affirming),
+            mapping.constrained.saturating_add(self.constrained),
+        )
+    }
+}
+
+/// Aligned positions where both members carry a literal, as
+/// `(frontier index, key pair)`.
+fn literal_positions(
+    canonical: &MemberContent,
+    member: &MemberContent,
+) -> Vec<(usize, (u64, u64))> {
+    canonical
+        .keys
+        .iter()
+        .zip(member.keys.iter())
+        .enumerate()
+        .filter(|(_, (left, right))| {
+            left.population == Population::Literal && right.population == Population::Literal
+        })
+        .map(|(index, (left, right))| (index, (left.key, right.key)))
+        .collect()
+}
+
+/// The key pairs of [`literal_positions`], for the literal bijection.
+fn literal_pairs(positions: &[(usize, (u64, u64))]) -> Vec<(u64, u64)> {
+    positions.iter().map(|(_, keys)| *keys).collect()
 }
 
 /// Rename-mapping evidence over one pair's aligned identifier positions
@@ -149,6 +256,32 @@ struct RenameMapping {
     /// The explained positions whose raw bytes are equal on both sides —
     /// evidence the substitution did not supply.
     identity: usize,
+}
+
+impl RenameMapping {
+    /// Whether the pair claims a rename at all: some substituted
+    /// identifier position is explained, so a bijection is asserting
+    /// that this copy was renamed rather than merely reused.
+    fn renames(&self) -> bool {
+        self.explained > self.identity
+    }
+
+    /// The identifier positions that anchor the proof.
+    ///
+    /// A window carved from inside a function that carries no literal at
+    /// all offers the substitution nothing to contradict — the literal
+    /// that would is on the line the window left out — so a substitution
+    /// corroborated only by its own repetition cannot anchor it. Its
+    /// anchors are the positions the rename did not supply: identity
+    /// identifiers ([FUSED-CONTENT-GATE-INTERIOR]). A whole authored
+    /// function or module with no literal is judged as before.
+    fn anchors(&self, scope: PairScope, aligned_literals: usize) -> usize {
+        if scope.interior && aligned_literals == 0 {
+            self.identity
+        } else {
+            self.explained
+        }
+    }
 }
 
 /// Measures [`RenameMapping`] for one pair's identifier positions,
@@ -204,7 +337,7 @@ fn rename_mapping(
 /// The aligned positions whose raw bytes differ — [TECH-PMATCH-BAKER]'s
 /// parameter alphabet, the population [`rename_mapping`] derives its
 /// bijection over.
-fn substituted_pairs(identifiers: &[(u64, u64)]) -> Vec<(u64, u64)> {
+pub(super) fn substituted_pairs(identifiers: &[(u64, u64)]) -> Vec<(u64, u64)> {
     identifiers
         .iter()
         .filter(|(left, right)| left != right)
@@ -264,199 +397,6 @@ fn evidence_weight(consistency: f64, anchors: usize) -> f64 {
         return 1.0;
     }
     weight
-}
-
-/// Literal echoes of the bijection's identifier substitutions (#409), as a
-/// per-substitution count: an aligned literal position whose bytes
-/// transform into the partner's bytes exactly by one bijection-explained
-/// identifier substitution. The transform is byte-exact replacement of
-/// every occurrence — content measurement over the leaf's raw bytes,
-/// the same bytes the keys hash — so `"OrderService"` echoes the
-/// `OrderService -> UserService` symbol substitution while a data
-/// table's `"GET"` against `"POST"` echoes nothing.
-fn literal_echoes<S: BuildHasher>(
-    canonical: &MemberContent,
-    member: &MemberContent,
-    sources: &HashMap<FileId, Vec<u8>, S>,
-) -> LiteralEchoes {
-    let identifiers = population(&canonical.keys, &member.keys, Population::Identifier);
-    let bijection = ModalBijection::over(&substituted_pairs(&identifiers));
-    let substitutions = explained_substitution_bytes(canonical, member, &bijection, sources);
-    let mut echoes = LiteralEchoes::default();
-    for index in substituted_literal_positions(canonical, member) {
-        let bytes = leaf_bytes(canonical, index, sources).zip(leaf_bytes(member, index, sources));
-        let Some((left, right)) = bytes else {
-            continue;
-        };
-        let explained_by = substitutions
-            .iter()
-            .find(|(_, (from, to))| replaced_matches(left, from, to, right));
-        if let Some((keys, _)) = explained_by {
-            let slot = echoes.per_substitution.entry(*keys).or_insert(0_usize);
-            *slot = slot.saturating_add(1);
-            let _newly = echoes.positions.insert(index);
-        }
-    }
-    echoes
-}
-
-/// Aligned literal positions that affirm the copy: positions whose raw
-/// bytes are preserved or whose bytes an echo explains. Every collapsed
-/// literal position counts on its own, the fragments of an interpolated
-/// string included — the frontier is positional, and
-/// [FUSED-CONTENT-GATE] pools each aligned literal position into the
-/// same coverage as the identifier positions. A preserved fragment is a
-/// preserved literal; the drifted fragment beside it is a drifted one,
-/// and weakens the proof in proportion like any other.
-fn affirming_literal_count(
-    canonical: &MemberContent,
-    member: &MemberContent,
-    echoes: &LiteralEchoes,
-) -> usize {
-    canonical
-        .keys
-        .iter()
-        .zip(member.keys.iter())
-        .enumerate()
-        .filter(|(_, (left, right))| {
-            left.population == Population::Literal && right.population == Population::Literal
-        })
-        .filter(|(index, (left, right))| left.key == right.key || echoes.positions.contains(index))
-        .count()
-}
-
-/// The echo evidence of one pair (#409): per-substitution counts for
-/// mapping corroboration, plus the frontier positions the echoes
-/// affirmed, for the authored-literal group discipline.
-#[derive(Default)]
-struct LiteralEchoes {
-    /// Echo count per bijection-explained substitution.
-    per_substitution: BTreeMap<(u64, u64), usize>,
-    /// Frontier indices whose literal bytes an echo explained.
-    positions: std::collections::BTreeSet<usize>,
-}
-
-/// Frontier indices of aligned positions where both members carry a
-/// literal and the raw bytes differ — the candidates an echo can
-/// explain.
-fn substituted_literal_positions(canonical: &MemberContent, member: &MemberContent) -> Vec<usize> {
-    canonical
-        .keys
-        .iter()
-        .zip(member.keys.iter())
-        .enumerate()
-        .filter(|(_, (left, right))| {
-            left.population == Population::Literal
-                && right.population == Population::Literal
-                && left.key != right.key
-        })
-        .map(|(index, _)| index)
-        .collect()
-}
-
-/// One bijection-explained substitution: the aligned key pair plus the
-/// raw bytes on each side.
-type SubstitutionBytes<'src> = ((u64, u64), (&'src [u8], &'src [u8]));
-
-/// The distinct bijection-explained identifier substitutions of one
-/// pair, with the raw bytes on each side — the substitution vocabulary
-/// [`literal_echoes`] tests candidates against.
-fn explained_substitution_bytes<'src, S: BuildHasher>(
-    canonical: &MemberContent,
-    member: &MemberContent,
-    bijection: &ModalBijection,
-    sources: &'src HashMap<FileId, Vec<u8>, S>,
-) -> Vec<SubstitutionBytes<'src>> {
-    let mut out: Vec<SubstitutionBytes<'src>> = Vec::new();
-    for (index, (left, right)) in canonical.keys.iter().zip(member.keys.iter()).enumerate() {
-        let keys = (left.key, right.key);
-        if left.population != Population::Identifier
-            || right.population != Population::Identifier
-            || left.key == right.key
-            || !bijection.explains(&keys)
-        {
-            continue;
-        }
-        if out.iter().any(|(seen, _)| *seen == keys) {
-            continue;
-        }
-        let bytes = leaf_bytes(canonical, index, sources).zip(leaf_bytes(member, index, sources));
-        if let Some(pair_bytes) = bytes {
-            out.push((keys, pair_bytes));
-        }
-    }
-    out
-}
-
-/// True when replacing the *symbol-boundary* occurrences of `from` in
-/// `left` with `to` yields exactly `right`, with at least one occurrence
-/// replaced. Pure byte-content equality under one substitution — no
-/// pattern language, no tokenisation; the leaves being compared were
-/// already isolated by the AST.
-///
-/// Replacing every raw byte occurrence instead accepted arbitrary data
-/// as rename proof: under an explained `a -> x` substitution, the literal
-/// `"banana"` transforms into `"bxnxnx"`, so a string whose payload
-/// merely *contains* the substituted bytes corroborated the rename it
-/// contradicts. Repeated across enough identifier positions that cleared
-/// [`CONTENT_SUPPORT_FLOOR`], it certified `rename_consistency = 1.0`
-/// for code whose literal data had changed. An echo is a *symbol* echo:
-/// the bytes have to occupy a place a symbol reference could occupy —
-/// `"OrderService"`, a name inside a path or a message — never the
-/// inside of a longer word ([REPAIR-RENAME-LITERAL-ECHO], gh #409).
-fn replaced_matches(left: &[u8], from: &[u8], to: &[u8], right: &[u8]) -> bool {
-    let mut expected: Vec<u8> = Vec::with_capacity(right.len());
-    let mut cursor = 0_usize;
-    let mut replaced = false;
-    while let Some(start) = next_occurrence(left, from, cursor) {
-        let Some(head) = left.get(cursor..start) else {
-            break;
-        };
-        expected.extend_from_slice(head);
-        let boundary = at_symbol_boundary(left, start, from.len());
-        expected.extend_from_slice(if boundary { to } else { from });
-        replaced = replaced || boundary;
-        cursor = start.saturating_add(from.len());
-    }
-    expected.extend_from_slice(left.get(cursor..).unwrap_or_default());
-    replaced && expected == right
-}
-
-/// First offset at or after `from_index` where `needle` occurs in
-/// `haystack`, `None` when there is none left.
-fn next_occurrence(haystack: &[u8], needle: &[u8], from_index: usize) -> Option<usize> {
-    let offset = find_bytes(haystack.get(from_index..)?, needle)?;
-    Some(from_index.saturating_add(offset))
-}
-
-/// True when the window `[start, start + len)` is delimited on both
-/// sides by a byte that cannot continue an identifier — the only place
-/// inside a literal payload where a symbol *reference* can sit. The
-/// quote characters that bound a string leaf count as delimiters, so a
-/// literal that is exactly the renamed symbol still echoes it.
-fn at_symbol_boundary(bytes: &[u8], start: usize, len: usize) -> bool {
-    let before = start.checked_sub(1).and_then(|index| bytes.get(index));
-    let after = bytes.get(start.saturating_add(len));
-    !before.is_some_and(|byte| is_word_byte(*byte))
-        && !after.is_some_and(|byte| is_word_byte(*byte))
-}
-
-/// True for a byte that continues an identifier-like word: ASCII
-/// alphanumerics and `_`, plus every non-ASCII byte, since a UTF-8 word
-/// continues through its lead and continuation bytes.
-fn is_word_byte(byte: u8) -> bool {
-    byte.is_ascii_alphanumeric() || byte == b'_' || !byte.is_ascii()
-}
-
-/// First byte offset of `needle` in `haystack`, `None` when absent or
-/// empty.
-fn find_bytes(haystack: &[u8], needle: &[u8]) -> Option<usize> {
-    if needle.is_empty() || haystack.len() < needle.len() {
-        return None;
-    }
-    haystack
-        .windows(needle.len())
-        .position(|window| window == needle)
 }
 
 /// The bidirectionally-modal substitution test shared by the substance
