+++ docs/specs/noise.md
@@ -189,7 +189,7 @@ This rule is the deliberate exception to the post-closure placement of the other
 The guard applies only when that exact pair needs embedding corroboration to pass admission: its structural and token evidence do not independently clear their route, while `E(p)` clears the configured embedding support floor. It reads the endpoint-keyed pair record directly and neither names nor inspects a component. Pair classification may describe the rejected pair for an explicit comparison, but no bucket, label, score, or evidence from this decision is copied to a cluster.
 
 ### [CLONE-NOISE-LITERAL-VARIATION-CALLS] Literal-variation call scaffolding
-Scaffolding repeats one call shape varying only its string-literal arguments — `setenv` keys, event names, endpoint paths — so after literal normalisation the members collapse to one subtree even though the differing literals are payload, not extractable logic. A cluster is suppressed when every member resolves to the same callee and arity (one enclosing call per member, or the same ordered call sequence contained in each member's range), at least one literal-bearing call position differs in string-literal bytes, and every literal-bearing position differs. A call position that carries no string literal is neutral: it neither proves variation nor blocks the sequence. An invariant literal-bearing position is shared authored logic and blocks suppression. Members whose literals all agree never match, so byte-identical copies keep the family's verbatim escape hatch.
+Scaffolding repeats one call shape varying only its string-literal arguments — `setenv` keys, event names, endpoint paths — so after literal normalisation the members collapse to one subtree even though the differing literals are payload, not extractable logic. A cluster is suppressed when every member resolves to the same callee and arity (one enclosing call per member, or the same ordered call sequence contained in each member's range), at least one literal-bearing call position differs in string-literal bytes, and every literal-bearing position differs. A call position that carries no string literal is neutral: it neither proves variation nor blocks the sequence, provided its bound result reaches a later varying position — a call **consumes** both its arguments and its receiver, so `expect(generated).toContain("…")` consumes `generated` (gh #284). An invariant literal-bearing position is shared authored logic and blocks suppression. Members whose literals all agree never match, so byte-identical copies keep the family's verbatim escape hatch.
 
 #### [CLONE-NOISE-LITERAL-VARIATION-CALLS-COVERED-STATEMENT] The covered-statement precondition
